Getting a start on my new Zilch dice game in C#
This will be an interactive dice game via a Windows Forms Application
written in C#. I used to play this game as a kid and have always wanted
to make it into a interactive PC App. It is fully compatible with
Windows PC touch screens. I may port this to the Windows 8 App Store,
and perhaps to some other mobile platforms so I am asserting a Copyright
© on all code that has been and will be written in the design and
development of this project. Darren Nelson Moody