Working on getting score btns working
The buttons seem to be popping up correctly, and I know how to fix them
so that they will auto-select the correct dice in a score sequence but
I've got to get some sleep tonight. And the wifey is getting annoyed at
the keyboard clicking. I'm really taking off on this project. C# is like
riding a bike...